Add authorisation for participants

+++ app/controllers/participants_controller.rb
@@ -1,9 +1,8 @@
 class ParticipantsController < ApplicationController
   before_filter :authenticate_user!
-  # after_action :verify_authorized
+  after_action :verify_authorized
 
   # TODO: authorization
-  # TODO: scope this model to devise -> invite_token?
   # TODO: Interface for participants -> seperated from the admin interface
 
 
@@ -11,16 +10,19 @@ class ParticipantsController < ApplicationController
 
   def index
     @participants = Participant.all
+    authorize Participant
     respond_with(@participants)
   end
 
   def show
     @participant = Participant.find(params[:id])
+    authorize @participant
     respond_with(@participant)
   end
 
   def new
     @participant = Participant.new
+    authorize @participant
     respond_with(@participant)
   end
 
@@ -28,16 +30,24 @@ class ParticipantsController < ApplicationController
     @participant = Participant.new(secure_params)
     @participant.password = SecureRandom.base64(69)
     @participant.save
-    @participant.send_reset_password_instructions
-    redirect_to participants_path, :notice => "Participant successfully created. An invitation email has been sent to #{@participant.email}."
+    authorize @participant
+
+    unless @participant.email.blank?
+      @participant.send_reset_password_instructions
+      redirect_to participants_path, :notice => "Participant successfully created. An invitation email has been sent to #{@participant.email}."
+    else
+      redirect_to participants_path, :notice => "Participant successfully created. (The invitation email was not sent because no email address was specified."
+    end
   end
 
   def edit
     @participant = Participant.find(params[:id])
+    authorize @participant
   end
 
   def update
     @participant = Participant.find(params[:id])
+    authorize @participant
     if @participant.update_attributes(secure_params)
       redirect_to participants_path, :notice => "Participant updated."
     else
@@ -47,12 +57,14 @@ class ParticipantsController < ApplicationController
 
   def destroy
     participant = Participant.find(params[:id])
+    authorize participant
     participant.destroy
     redirect_to users_path, :notice => "Participant deleted."
   end
 
   def resend_invite
     @participant = Participant.find(params[:id])
+    authorize @participant
     if @participant
       if !@participant.email.blank?
         @participant.send_reset_password_instructions
